fix: using size_t instead of int for GeoElementId indices

+++ src/Mod/Sketcher/App/Constraint.cpp
@@ -478,7 +478,7 @@ void Constraint::addElement(GeoElementId element)
 #endif
 }
 
-int Constraint::getGeoId(int index) const
+int Constraint::getGeoId(size_t index) const
 {
 #if SKETCHER_CONSTRAINT_USE_LEGACY_ELEMENTS
     if (index < 3) {
@@ -495,7 +495,7 @@ int Constraint::getGeoId(int index) const
     return hasElement(index) ? elements[index].GeoId : GeoEnum::GeoUndef;
 }
 
-PointPos Constraint::getPosId(int index) const
+PointPos Constraint::getPosId(size_t index) const
 {
 #if SKETCHER_CONSTRAINT_USE_LEGACY_ELEMENTS
     if (index < 3) {
@@ -512,7 +512,7 @@ PointPos Constraint::getPosId(int index) const
     return hasElement(index) ? elements[index].Pos : PointPos::none;
 }
 
-int Constraint::getPosIdAsInt(int index) const
+int Constraint::getPosIdAsInt(size_t index) const
 {
 #if SKETCHER_CONSTRAINT_USE_LEGACY_ELEMENTS
     if (index < 3) {
@@ -529,7 +529,7 @@ int Constraint::getPosIdAsInt(int index) const
     return hasElement(index) ? elements[index].posIdAsInt() : 0;
 }
 
-bool Constraint::hasElement(int index) const
+bool Constraint::hasElement(size_t index) const
 {
     return index >= 0 && static_cast<decltype(elements)::size_type>(index) < elements.size();
 }
@@ -554,7 +554,7 @@ size_t Constraint::getElementIndexForGeoId(int geoId) const
     return -1;
 }
 
-void Constraint::setGeoId(int index, int geoId)
+void Constraint::setGeoId(size_t index, int geoId)
 {
 #if SKETCHER_CONSTRAINT_USE_LEGACY_ELEMENTS
     if (index < 3) {
@@ -576,7 +576,7 @@ void Constraint::setGeoId(int index, int geoId)
     }
 }
 
-void Constraint::setPosId(int index, PointPos pos)
+void Constraint::setPosId(size_t index, PointPos pos)
 {
 #if SKETCHER_CONSTRAINT_USE_LEGACY_ELEMENTS
     if (index < 3) {
@@ -598,7 +598,7 @@ void Constraint::setPosId(int index, PointPos pos)
     }
 }
 
-void Constraint::setPosId(int index, int pos)
+void Constraint::setPosId(size_t index, int pos)
 {
 #if SKETCHER_CONSTRAINT_USE_LEGACY_ELEMENTS
     if (index < 3) {
@@ -620,7 +620,7 @@ void Constraint::setPosId(int index, int pos)
     }
 }
 
-bool Constraint::ensureElementExists(int index)
+bool Constraint::ensureElementExists(size_t index)
 {
     if (index < 0) {
         return false;  // Indicate failure for an invalid index
@@ -631,7 +631,7 @@ bool Constraint::ensureElementExists(int index)
     return true;
 }
 
-void Constraint::swapElements(int index1, int index2)
+void Constraint::swapElements(size_t index1, size_t index2)
 {
     if (index1 == index2) {
         return;
